requirements.txt
robotframework
suds-py3